Rename logging to not conflict with python internal logging.
