+++ tests/integration/library/cockroach/cockroach_test_case.py
@@ -5,7 +5,7 @@ from unittest import SkipTest
 from discograph import database
 from discograph.library.cache.cache_manager import setup_cache, shutdown_cache
 from discograph.config import CockroachTestConfiguration
-from discograph.logging import setup_logging, shutdown_logging
+from discograph.logging_config import setup_logging, shutdown_logging
 
 log = logging.getLogger(__name__)
