+++ tests/integration/library/sqlite/sqlite_test_case.py
@@ -4,7 +4,7 @@ import unittest
 from discograph import database
 from discograph.library.cache.cache_manager import setup_cache, shutdown_cache
 from discograph.config import SqliteTestConfiguration
-from discograph.logging import setup_logging, shutdown_logging
+from discograph.logging_config import setup_logging, shutdown_logging
 
 log = logging.getLogger(__name__)
